fix issues with the discovery feature
If mprisqueeze runs behing a firewall, as no answer can arrive from the
LMS server, it will flood the network with UDP broadcast messages
causing many issues.

This is resolved by sending the UDP broadcast message and waiting for the
answer synchronously.

Somehow related, squeelite also try to discover the LMS server using the
same technic but not using connected mode. This raises another issue
with the firewall. iptables does not recognize a connected state between
a broadcast message and its response.

This is resolved by passing the IP of the server to squeelite.